requirements.txt
pillow
python-docx
pytesseract
